+++ test/test_compute_vectors_integration.py
@@ -2,7 +2,6 @@
 End -> End Testing of the vector computation and sorting.
 """
 
-from pathlib import Path
 from test.assets import (
     SORTED_BENCH_SCENES_PATHS,
     SORTED_STREAM_GENERIC_PATHS,
@@ -13,15 +12,21 @@ from typing import List
 import pytest
 
 import content_aware_timelapse.frames_to_vectors.conversion
-from content_aware_timelapse import vector_scoring
-from content_aware_timelapse.frames_to_vectors.vector_computation.compute_vectors_vit import (
-    compute_vectors_vit_cls,
+import content_aware_timelapse.frames_to_vectors.vector_computation.compute_vectors_vit
+from content_aware_timelapse.frames_to_vectors import vector_scoring
+from content_aware_timelapse.frames_to_vectors.conversion_types import ConversionScoringFunctions
+from content_aware_timelapse.frames_to_vectors.vector_computation.compute_vectors_clip import (
+    CONVERT_CLIP,
 )
-from content_aware_timelapse.vector_scoring import IndexScores
+from content_aware_timelapse.frames_to_vectors.vector_computation.compute_vectors_vit import (
+    CONVERT_VIT_CLS,
+)
+from content_aware_timelapse.frames_to_vectors.vector_scoring import IndexScores
 from content_aware_timelapse.viderator import image_common
 from content_aware_timelapse.viderator.image_common import ImageSourceType
 
 
+@pytest.mark.parametrize("conversion_scoring_functions", [CONVERT_CLIP, CONVERT_VIT_CLS])
 @pytest.mark.parametrize(
     "frames",
     [
@@ -30,7 +35,9 @@ from content_aware_timelapse.viderator.image_common import ImageSourceType
         list(map(image_common.load_rgb_image, SORTED_STREAM_GENERIC_PATHS)),
     ],
 )
-def test_sorting_converted_vectors(frames: ImageSourceType) -> None:
+def test_sorting_converted_vectors(
+    conversion_scoring_functions: ConversionScoringFunctions, frames: ImageSourceType
+) -> None:
     """
     Sanity check using library assets to assume that scoring and sorting images works as expected.
     :param frames: Should be sorted in order of interesting-ness, with the most interesting frame
@@ -43,18 +50,21 @@ def test_sorting_converted_vectors(frames: ImageSourceType) -> None:
         input_signature="test signature",
         batch_size=1,
         total_input_frames=1,
-        convert_batches=compute_vectors_vit_cls,
+        convert_batches=conversion_scoring_functions.conversion,
         intermediate_path=None,
     )
 
     score_indexes: List[IndexScores] = list(
-        map(vector_scoring.calculate_scores, enumerate(vectors))
+        map(
+            conversion_scoring_functions.scoring,
+            enumerate(vectors),
+        )
     )
 
     indices = vector_scoring._score_and_sort_frames(  # pylint: disable=protected-access
         index_scores=score_indexes,
         num_output_frames=len(score_indexes) - 2,
-        plot_path=Path("./fig.png"),
+        plot_path=None,
     )
 
     assert list(sorted(indices, reverse=True)) == indices